requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
google-cloud-aiplatform>=1.38.1
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.16.0
opencv-python>=4.8.1.78
numpy>=1.26.0
python-dotenv>=1.0.0
google-generativeai>=0.1.0
vertexai>=0.1.0
google-genai>=0.1.0
bcrypt>=4.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
